requirements.txt
# Web Framework
streamlit==1.29.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# AI Integration
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2

# File handling
python-magic==0.4.27
